Split Population classes into their own module
