+++ tests/core/test_countries.py
@@ -7,7 +7,7 @@ from openfisca_core.periods import MONTH
 from openfisca_core.simulation_builder import SimulationBuilder
 from openfisca_core.taxbenefitsystems import VariableNameConflict, VariableNotFound
 from openfisca_core import periods
-from openfisca_core.entities import DIVIDE
+from openfisca_core.populations import DIVIDE
 from openfisca_country_template import CountryTaxBenefitSystem
 from openfisca_core.tools import assert_near
