+++ tests/core/test_simulation_builder.py
@@ -10,7 +10,8 @@ from pytest import raises, fixture, approx
 from openfisca_core.simulation_builder import SimulationBuilder, Simulation
 from openfisca_core.tools import assert_near
 from openfisca_core.tools.test_runner import yaml
-from openfisca_core.entities import Entity, GroupEntity, Population
+from openfisca_core.entities import Entity, GroupEntity
+from openfisca_core.populations import Population
 from openfisca_core.variables import Variable
 from openfisca_country_template.entities import Household
 from openfisca_country_template.situation_examples import couple
